# back-end/schema/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Personal trading account API
servers:
  - url: http://localhost:3001/api
paths:
...
  /site/createSite:
    post:
      tags:
        - site
      summary: Create site
      description: Create site (only with permissions admin or superAdmin)
      operationId: createSite
      requestBody:
        description: Site object that needs to be created
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: integer
                  example: 'Site name'
                url:
                  type: integer
                  example: 'https://www.google.com'
                logo:
                  type: string
                  format: binary
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site created'
        '400':
          description: Site not created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site not created'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
...
  /site/getSites/{siteId}:
    get:
      tags:
        - site
      summary: Get site by id
      description: Get site by id
      operationId: getSite
      parameters:
        - name: siteId
          in: path
          description: ID of site to return
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      security: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site found'
                  data:
                    type: object
                    $ref: '#/definitions/Site'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site not found'
    delete:
      tags:
        - site
      summary: Delete site by id
      description: Delete site by id (only with permissions admin or superAdmin)
      operationId: deleteSite
      parameters:
        - name: siteId
          in: path
          description: ID of site to delete
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site deleted'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site not found'
    put:
      tags:
        - site
      summary: Update site by id
      description: Update site by id (only with permissions admin or superAdmin)
      operationId: updateSite
      parameters:
        - name: siteId
          in: path
          description: ID of site to update
          required: true
          schema:
            type: string
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      requestBody:
        description: Site object that needs to be updated
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: integer
                  example: 'Site name'
                url:
                  type: integer
                  example: 'https://www.google.com'
                logo:
                  type: string
                  format: binary
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site updated'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site not found'
  /site/getSites/{siteId}/createHistory:
    post:
      tags:
        - site
      summary: Create history for site
      description: Create history for site
      operationId: createHistory
      parameters:
        - name: siteId
          in: path
          description: ID of site to create history
          required: true
          schema:
            type: string
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      requestBody:
        description: History object that needs to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/History'
            example:
              transactions: ['Some transaction']
              purchases: ['Some purchase']
              balance: 100
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'History created'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Site not found'
...
  /site/getSites/{siteId}/histories/{historyId}:
    get:
      tags:
        - site
      summary: Get history by id
      description: Get history by id
      operationId: getHistoryById
      parameters:
        - name: siteId
          in: path
          description: ID of site to get history
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
        - name: historyId
          in: path
          description: ID of history to return
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'History found'
                  data:
                    type: object
                    $ref: '#/definitions/History'
        '404':
          description: History not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'History not found'
    put:
      tags:
        - site
      summary: Update history by id
      description: Update history by id
      operationId: updateHistory
      parameters:
        - name: siteId
          in: path
          description: ID of site to update history
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
        - name: historyId
          in: path
          description: ID of history to update
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      requestBody:
        description: History object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/History'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'History updated'
        '404':
          description: History not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'History not found'
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
security:
  - cookieAuth: []
definitions:
...
  Site:
    required:
      - name
      - url
      - type
    type: object
    properties:
      _id:
        type: ObjectId
        format: _id
        example: 63080201e3957808e2bebe61
      type:
        type: string
        default: Site
      name:
        type: string
        example: Site1
      url:
        type: string
        example: https://site1.com
      logo:
        type: string
        example: https://site1.com/logo.png
      created:
        type: number
        default: 1663117006355
      updated:
        type: number
        default: 1663117006355
    xml:
      name: site
  History:
    required:
      - siteId
    type: object
    properties:
      _id:
        type: ObjectId
        format: _id
        example: 63080201e3957808e2bebe61
      type:
        type: string
        default: History
      siteId:
        type: ObjectId
        ref: Site
        example: 63080201e3957808e2bebe61
      userId:
        type: ObjectId
        ref: User
        example: 63080201e3957808e2bebe61
      balance:
        type: number
        default: 0
      transactions:
        type: array
        items:
          type: object
      purchases:
        type: array
        items:
          type: object
      created:
        type: number
        default: 1663117006355
      updated:
        type: number
        default: 1663117006355
    xml:
      name: history
...